Rename secret_key to p2p_secret_key
Also make the implementation in `stage` and `p2p` the same for this
option.
